ui design updates

+++ app/src/main/java/com/petko/fruitapp/utils/Fruit.kt
@@ -17,6 +17,6 @@ data class Nutrition(
     val carbohydrates: Double,
     val protein: Double,
     val fat: Double,
-    val calories: Int,
+    val calories: Double,
     val sugar: Double,
 )
